+++ .gitignore
@@ -14,3 +14,4 @@ package-lock.json
 example-plugins
 
 .idea
+dist
