+++ bin/homebridge
@@ -9,9 +9,9 @@
 process.title = 'homebridge';
 
 // Find the HomeBridge lib
-var path = require('path');
-var fs = require('fs');
-var lib = path.join(path.dirname(fs.realpathSync(__filename)), '../lib');
+const path = require('path');
+const fs = require('fs');
+const dist = path.join(path.dirname(fs.realpathSync(__filename)), '../dist');
 
 // Run HomeBridge
-require(lib + '/cli')();
+require(dist + '/cli').default();
