+++ lib/types/node-persist.d.ts
@@ -1,0 +1,3 @@
+declare module 'node-persist' {
+  export function create(): any;
+}
