+++ lib/types/qrcode-terminal.d.ts
@@ -1,0 +1,3 @@
+declare module 'qrcode-terminal' {
+  export function generate(data: string): void;
+}
