+++ package.json
@@ -3,6 +3,7 @@
   "description": "HomeKit support for the impatient",
   "version": "0.4.50",
   "scripts": {
+    "build": "rimraf dist/ && tsc",
     "dev": "DEBUG=* ./bin/homebridge -D -P example-plugins/ || true"
   },
   "author": {
@@ -25,11 +26,17 @@
   },
   "preferGlobal": true,
   "dependencies": {
-    "chalk": "^1.1.1",
-    "commander": "2.8.1",
-    "hap-nodejs": "0.4.51",
-    "semver": "5.0.3",
+    "chalk": "^2.4.2",
+    "commander": "^2.20.0",
+    "hap-nodejs": "0.4.52",
     "node-persist": "^0.0.8",
-    "qrcode-terminal": "^0.11.0"
+    "qrcode-terminal": "^0.11.0",
+    "semver": "5.0.3"
+  },
+  "devDependencies": {
+    "@types/node": "^12.0.10",
+    "@types/semver": "^6.0.1",
+    "rimraf": "^2.6.3",
+    "typescript": "^3.5.2"
   }
 }
